backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
watchdog==3.0.0
python-docx==0.8.11
pdfminer.six==20221105
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.7
reportlab==4.0.0
aiofiles==23.1.0
requests==2.31.0
numpy<2
huggingface-hub==0.23.0
transformers==4.41.2
torch==2.3.0
